+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 2025-12-24
+
+### Added
+- Server deployment configuration for AWS Fargate and Lambda
+
 ## 2025-12-23
 
 ### Added